chore: moved the borderbox file inside apps/tasklist
